A republish the project up to this point
Due to the mismanagement of a local config file, commits to this repo were pushed under the pseudonym "SamVimes22". This is a literary reference, and will be familiar to anyone who has perused a significant portion of Terry Prachett's Discworld. Rest assured that the work presented here is entirely my own; nobody else I know is crazy enough to try and build something like this.